@@ -7,15 +7,16 @@
     "dist"
   ],
   "scripts": {
-    "start": "npm run dev-watch & npm run compile & npm run test-watch",
+    "start": "npm run compile-watch & npm run test-watch",
     "dev": "webpack --mode development",
-    "dev-watch": "webpack --mode development --watch --hot",
     "prod": "webpack --mode production",
     "clean": "rimraf ./coverage && rimraf ./dist/tags",
-    "compile": "npm run clean && riot -w tags dist/tags",
-    "test": "npm run clean && riot tags dist/tags && karma start",
+    "compile": "npm run clean && riot tags --output dist/tags",
+    "compile-watch": "npm run clean && riot -w tags --output dist/tags",
+    "test": "karma start",
     "test-watch": "karma start --no-single-run",
-    "build": "npm run test && npm run dev && npm run prod"
+    "test-run": "karma run",
+    "build": "npm run compile && npm run test && npm run dev && npm run prod"
   },
   "keywords": [
     "riotjs",
@@ -38,9 +39,13 @@
   "dependencies": {
     "date-fns": "^1.29.0",
     "q": "^1.5.1",
-    "riot": "^3.13.2"
+    "riot": "4.6.6",
+    "riot-observable": "^3.0.0"
   },
   "devDependencies": {
+    "@riotjs/cli": "^4.0.2",
+    "@riotjs/compiler": "4.5.2",
+    "@riotjs/webpack-loader": "^4.0.1",
     "babel-core": "^6.25.0",
     "babel-loader": "^7.1.4",
     "babel-preset-es2015-riot": "^1.1.0",
@@ -51,7 +56,6 @@
     "htmlhint-loader": "^1.3.1",
     "htmlhint-riot-rules": "^0.3.14",
     "istanbul-instrumenter-loader": "^3.0.1",
-    "jquery": "^3.3.1",
     "karma": "^4.0.1",
     "karma-chrome-launcher": "^2.2.0",
     "karma-coverage-istanbul-reporter": "^2.0.4",
@@ -66,8 +70,9 @@
     "riotjs-loader": "^4.0.0",
     "sinon": "^2.1.0",
     "sinon-chai": "^2.13.0",
-    "webpack": "^4.8.3",
-    "webpack-cli": "^3.3.0",
+    "uuid": "^3.3.3",
+    "webpack": "^4.39.1",
+    "webpack-cli": "^3.3.6",
     "webpack-date-fns-externals": "^1.0.2"
   }
 }
