@@ -1,165 +1,205 @@
-const fireEvent = require('../../helpers').fireEvent
-require('../../../dist/tags/datepicker/su-datepicker.js')
-
-riot.mixin('semantic-ui', {})
+import * as riot from 'riot'
+import { init, compile } from '../../helpers/'
+import TargetComponent from '../../../dist/tags/datepicker/su-datepicker.js'
 
 describe('su-datepicker', function () {
-  let tag
+  let element, component
   let spyOnOpen = sinon.spy()
   let spyOnClose = sinon.spy()
   let spyOnClick = sinon.spy()
   let spyOnChange = sinon.spy()
-  let mount = opts => {
-    tag = riot.mount('su-datepicker', opts)[0]
-    tag.on('open', spyOnOpen)
-      .on('close', spyOnClose)
-      .on('click', spyOnClick)
-      .on('change', spyOnChange)
+  init(riot)
+
+  const mount = opts => {
+    const option = Object.assign({
+      'onopen': spyOnOpen,
+      'onclose': spyOnClose,
+      'onclick': spyOnClick,
+      'onchange': spyOnChange,
+    }, opts)
+    element = document.createElement('app')
+    riot.register('su-datepicker', TargetComponent)
+    const AppComponent = compile(`
+      <app>
+        <su-datepicker
+          onopen="{ () => dispatch('open') }"
+          onclose="{ () => dispatch('close') }"
+          onclick="{ () => dispatch('click') }"
+          onchange="{ () => dispatch('change') }"
+          popup="{ props.popup }"
+          current-date="{ props.currentDate }"
+          value="{ value || props.value }"
+          placeholder="{ props.placeholder }"
+          pattern="{ props.pattern }"
+          locale="{ props.locale }"
+          tabindex="{ props.tabindex }"
+          start-mode="{ props.startMode }"
+          year-range="{ props.yearRange }"
+          class="{ props.class }"
+          direction="{ props.direction }"
+        />
+        <button id="reset" type="button" onclick="{ reset }">reset</button>
+        <script>
+          export default {
+            reset() {
+              this.obs.trigger(this.$('su-datepicker').id + '-reset')
+            }
+          }
+        </script>
+      </app>`)
+    riot.register('app', AppComponent)
+    component = riot.mount(element, option)[0]
   }
+
   const addZero = num => {
     return ('0' + num).slice(-2);
   }
-
-  beforeEach(function () {
-  })
 
   afterEach(function () {
     spyOnOpen.reset()
     spyOnClose.reset()
     spyOnClick.reset()
     spyOnChange.reset()
-    tag.unmount()
+    riot.unregister('su-datepicker')
+    riot.unregister('app')
   })
 
   it('is mounted', function () {
-    $('body').append('<su-datepicker />')
     mount()
-    tag.isMounted.should.be.true
+    expect(component).to.be.ok
 
-    $('su-datepicker .ui.action.input').length.should.equal(0)
-    spyOnOpen.should.have.been.callCount(0)
-    spyOnClose.should.have.been.callCount(0)
+    expect(component.$('.ui.action.input')).to.be.undefined
+    expect(spyOnOpen).to.have.been.not.called
+    expect(spyOnClose).to.have.been.not.called
   })
 
   it('opens/closes datepicker and triggers open/close event', function () {
-    $('body').append('<su-datepicker />')
-    mount({ popup: true })
-    $('su-datepicker .menu').css('visibility').should.equal('hidden')
+    mount({
+      popup: true
+    })
+    expect(component.$('.menu').classList.contains('hidden')).to.be.ok
 
-    fireEvent($('su-datepicker button.ui.icon.button')[0], 'click')
-    spyOnOpen.should.have.been.calledOnce
-    $('su-datepicker .menu').css('visibility').should.equal('visible')
+    fireEvent(component.$('button.ui.icon.button'), 'click')
+    expect(spyOnOpen).to.have.been.calledOnce
+    expect(component.$('.menu').classList.contains('visible')).to.be.ok
 
-    fireEvent($('su-datepicker .dp-day .ui.button:first')[0], 'click')
-    spyOnClick.should.have.been.calledOnce
-    spyOnChange.should.have.been.calledOnce
-    spyOnClose.should.have.been.calledOnce
-    $('su-datepicker .menu').css('visibility').should.equal('hidden')
+    fireEvent(component.$('.dp-day .ui.button'), 'click')
+    expect(spyOnClick).to.have.been.calledOnce
+    expect(spyOnChange).to.have.been.calledOnce
+    expect(spyOnClose).to.have.been.calledOnce
+    expect(component.$('.menu').classList.contains('hidden')).to.be.ok
 
 
-    fireEvent($('su-datepicker button.ui.icon.button')[0], 'click')
-    spyOnOpen.should.have.been.calledTwice
-    $('su-datepicker .menu').css('visibility').should.equal('visible')
+    fireEvent(component.$('button.ui.icon.button'), 'click')
+    expect(spyOnOpen).to.have.been.calledTwice
+    expect(component.$('.menu').classList.contains('visible')).to.be.ok
 
-    fireEvent($('su-datepicker button.ui.icon.button')[0], 'click')
-    spyOnClose.should.have.been.calledTwice
-    $('su-datepicker .menu').css('visibility').should.equal('hidden')
+    fireEvent(component.$('button.ui.icon.button'), 'click')
+    expect(spyOnClose).to.have.been.calledTwice
+    expect(component.$('.menu').classList.contains('hidden')).to.be.ok
 
-    fireEvent($('su-datepicker button.ui.icon.button')[0], 'click')
-    spyOnOpen.should.have.been.callCount(3)
-    $('su-datepicker .menu').css('visibility').should.equal('visible')
+    fireEvent(component.$('button.ui.icon.button'), 'click')
+    expect(spyOnOpen).to.have.been.callCount(3)
+    expect(component.$('.menu').classList.contains('visible')).to.be.ok
 
-    fireEvent($('su-datepicker .menu')[0], 'mousedown')
-    fireEvent($('su-datepicker button.ui.icon.button')[0], 'blur')
-    fireEvent($('su-datepicker .menu')[0], 'mouseup')
-    $('su-datepicker .menu').css('visibility').should.equal('visible')
-    spyOnClose.should.have.been.calledTwice
+    fireEvent(component.$('.menu'), 'mousedown')
+    fireEvent(component.$('button.ui.icon.button'), 'blur')
+    fireEvent(component.$('.menu'), 'mouseup')
+    expect(component.$('.menu').classList.contains('visible')).to.be.ok
+    expect(spyOnClose).to.have.been.calledTwice
 
-    fireEvent($('su-datepicker .menu')[0], 'blur')
-    $('su-datepicker .menu').css('visibility').should.equal('hidden')
-    spyOnClose.should.have.been.callCount(3)
+    fireEvent(component.$('.menu'), 'blur')
+    expect(component.$('.menu').classList.contains('hidden')).to.be.ok
+    expect(spyOnClose).to.have.been.callCount(3)
   })
 
   it('select month event', function () {
-    $('body').append('<su-datepicker />')
-    mount({ currentDate: new Date(2017, 11, 1) })
-    tag.currentDate.getTime().should.equal(new Date(2017, 11, 1).getTime())
+    mount({
+      currentDate: new Date(2017, 11, 1)
+    })
+    expect(component.$('.dp-navigation .year').innerText).to.equal('2017')
+    expect(component.$('.dp-navigation .month').innerText).to.equal('Dec')
 
-    fireEvent($('su-datepicker .dp-navigation .month')[0], 'click')
-    $('su-datepicker .dp-day').length.should.equal(0)
-    $('su-datepicker .dp-month').length.should.equal(12)
-    $('su-datepicker .dp-month:first').text().trim().should.equal('Jan')
+    fireEvent(component.$('.dp-navigation .month'), 'click')
+    expect(component.$('.dp-day')).to.be.undefined
+    expect(component.$$('.dp-month').length).to.equal(12)
+    expect(component.$('.dp-month').innerText.trim()).to.equal('Jan')
 
-    fireEvent($('su-datepicker .dp-month:first .button')[0], 'click')
-    tag.currentDate.getTime().should.equal(new Date(2017, 0, 1).getTime())
+    fireEvent(component.$('.dp-month .button'), 'click')
+    expect(component.$('.dp-navigation .year').innerText).to.equal('2017')
+    expect(component.$('.dp-navigation .month').innerText).to.equal('Jan')
 
-    $('su-datepicker .dp-day').length.should.equal(7 * 6)
-    $('su-datepicker .dp-month').length.should.equal(0)
+    expect(component.$$('.dp-day').length).to.equal(7 * 6)
+    expect(component.$('.dp-month')).to.be.undefined
   })
 
   it('next/prev month event', function () {
-    $('body').append('<su-datepicker />')
-    mount({ currentDate: new Date(2017, 11, 1) })
-    tag.currentDate.getTime().should.equal(new Date(2017, 11, 1).getTime())
+    mount({
+      currentDate: new Date(2017, 11, 1)
+    })
+    expect(component.$('.dp-day').innerText).to.be.equal('26')
+    expect(component.$('.dp-navigation .year').innerText).to.equal('2017')
+    expect(component.$('.dp-navigation .month').innerText).to.equal('Dec')
 
-    fireEvent($('su-datepicker .dp-navigation .prev')[0], 'click')
-    $('su-datepicker .dp-day').length.should.equal(7 * 6)
-    $('su-datepicker .dp-month').length.should.equal(0)
-    $('su-datepicker .dp-day:first').text().trim().should.equal('29')
-    tag.currentDate.getTime().should.equal(new Date(2017, 10, 1).getTime())
+    fireEvent(component.$('.dp-navigation .prev'), 'click')
+    expect(component.$$('.dp-day').length).to.be.equal(7 * 6)
+    expect(component.$('.dp-month')).to.be.undefined
+    expect(component.$('.dp-day').innerText).to.be.equal('29')
+    expect(component.$('.dp-navigation .year').innerText).to.equal('2017')
+    expect(component.$('.dp-navigation .month').innerText).to.equal('Nov')
 
-    fireEvent($('su-datepicker .dp-navigation .next')[0], 'click')
-    $('su-datepicker .dp-day:first').text().trim().should.equal('26')
-    tag.currentDate.getTime().should.equal(new Date(2017, 11, 1).getTime())
+    fireEvent(component.$('.dp-navigation .next'), 'click')
+    expect(component.$('.dp-day').innerText).to.be.equal('26')
+    expect(component.$('.dp-navigation .year').innerText).to.equal('2017')
+    expect(component.$('.dp-navigation .month').innerText).to.equal('Dec')
   })
 
   it('select year event', function () {
-    $('body').append('<su-datepicker />')
-    mount({ currentDate: new Date(2017, 11, 1) })
-    tag.currentDate.getTime().should.equal(new Date(2017, 11, 1).getTime())
+    mount({
+      currentDate: new Date(2017, 11, 1)
+    })
+    expect(component.$('.dp-navigation .year').innerText).to.equal('2017')
+    expect(component.$('.dp-navigation .month').innerText).to.equal('Dec')
 
-    fireEvent($('su-datepicker .dp-navigation .year')[0], 'click')
-    $('su-datepicker .dp-day').length.should.equal(0)
-    $('su-datepicker .dp-month').length.should.equal(20)
-    $('su-datepicker .dp-month:first').text().trim().should.equal('2008')
+    fireEvent(component.$('.dp-navigation .year'), 'click')
+    expect(component.$('.dp-day')).to.be.undefined
+    expect(component.$$('.dp-month').length).to.be.equal(20)
+    expect(component.$('.dp-month').innerText).to.be.equal('2008')
 
-    fireEvent($('su-datepicker .dp-navigation .next')[0], 'click')
-    $('su-datepicker .dp-month:first').text().trim().should.equal('2028')
+    fireEvent(component.$('.dp-navigation .next'), 'click')
+    expect(component.$('.dp-month').innerText).to.be.equal('2028')
 
-    fireEvent($('su-datepicker .dp-navigation .prev')[0], 'click')
-    $('su-datepicker .dp-month:first').text().trim().should.equal('2008')
+    fireEvent(component.$('.dp-navigation .prev'), 'click')
+    expect(component.$('.dp-month').innerText).to.be.equal('2008')
 
-    fireEvent($('su-datepicker .dp-month:first .button')[0], 'click')
-    tag.currentDate.getTime().should.equal(new Date(2008, 11, 1).getTime())
+    fireEvent(component.$('.dp-month .button'), 'click')
 
-    $('su-datepicker .dp-day').length.should.equal(0)
-    $('su-datepicker .dp-month').length.should.equal(12)
+    expect(component.$('.dp-day')).to.be.undefined
+    expect(component.$$('.dp-month').length).to.be.equal(12)
   })
 
   it('popup datepicker', function () {
-    $('body').append('<su-datepicker />')
     mount({
       popup: true,
       currentDate: new Date(2017, 11, 1)
     })
-    $('su-datepicker .menu').css('visibility').should.equal('hidden')
+    expect(component.$('.menu').classList.contains('hidden')).to.be.ok
 
-    fireEvent($('su-datepicker button.ui.icon.button')[0], 'click')
-    spyOnOpen.should.have.been.calledOnce
-    $('su-datepicker .menu').css('visibility').should.equal('visible')
+    fireEvent(component.$('button.ui.icon.button'), 'click')
+    expect(spyOnOpen).to.have.been.calledOnce
+    expect(component.$('.menu').classList.contains('visible')).to.be.ok
 
-    fireEvent($('su-datepicker .dp-day .ui.button:eq(5)')[0], 'click')
-    spyOnClose.should.have.been.calledOnce
-    $('su-datepicker .menu').css('visibility').should.equal('hidden')
-    tag.refs.input.value.should.equal('2017-12-01')
+    fireEvent(component.$$('.dp-day .ui.button')[5], 'click')
+    expect(spyOnClose).to.have.been.calledOnce
+    expect(component.$('.menu').classList.contains('hidden')).to.be.ok
+    expect(component.$('input').value).to.equal('2017-12-01')
 
-    tag.value = new Date(2018, 11, 1)
-    tag.update()
-    tag.refs.input.value.should.equal('2018-12-01')
+    component.value = new Date(2018, 11, 1)
+    component.update()
+    expect(component.$('input').value).to.equal('2018-12-01')
   })
 
   it('popup datepicker option', function () {
-    $('body').append('<su-datepicker />')
     mount({
       popup: true,
       currentDate: new Date(2017, 11, 1),
@@ -168,302 +208,264 @@ describe('su-datepicker', function () {
       locale: require('date-fns/locale/ja'),
       tabindex: 10
     })
-    $('su-datepicker .menu').css('visibility').should.equal('hidden')
-    $('su-datepicker .dp-weekday:first').text().trim().should.equal('日')
-    $('su-datepicker .dp-navigation .month').text().trim().should.equal('12月')
+    expect(component.$('.menu').classList.contains('hidden')).to.be.ok
+    expect(component.$('.dp-weekday').innerText.trim()).to.equal('日')
+    expect(component.$('.dp-navigation .month').innerText.trim()).to.equal('12月')
 
-    fireEvent($('su-datepicker button.ui.icon.button')[0], 'click')
-    spyOnOpen.should.have.been.calledOnce
-    $('su-datepicker .menu').css('visibility').should.equal('visible')
-    $('su-datepicker input').attr('placeholder').should.equal('YYYY/MM/DD')
-    $('su-datepicker input').attr('tabindex').should.equal('10')
+    fireEvent(component.$('button.ui.icon.button'), 'click')
+    expect(spyOnOpen).to.have.been.calledOnce
+    expect(component.$('.menu').classList.contains('visible')).to.be.ok
+    expect(component.$('input').getAttribute('placeholder')).to.equal('YYYY/MM/DD')
+    expect(component.$('input').getAttribute('tabindex')).to.equal('10')
 
-    fireEvent($('su-datepicker .dp-day .ui.button:eq(5)')[0], 'click')
-    spyOnClose.should.have.been.calledOnce
-    $('su-datepicker .menu').css('visibility').should.equal('hidden')
-    tag.refs.input.value.should.equal('2017/12/01')
+    fireEvent(component.$$('.dp-day .ui.button')[5], 'click')
+    expect(spyOnClose).to.have.been.calledOnce
+    expect(component.$('.menu').classList.contains('hidden')).to.be.ok
+    expect(component.$('input').value).to.equal('2017/12/01')
   })
 
   it('popup datepicker default value', function () {
-    $('body').append('<su-datepicker />')
     mount({
       popup: true,
       pattern: 'YYYY/MM/DD',
       value: '2017/12/01',
     })
-    tag.refs.input.value.should.equal('2017/12/01')
+    expect(component.$('input').value).to.equal('2017/12/01')
   })
 
   it('year first option', function () {
-    $('body').append('<su-datepicker />')
     mount({ currentDate: new Date(2017, 11, 1), popup: true, startMode: "year" })
-    tag.currentDate.getTime().should.equal(new Date(2017, 11, 1).getTime())
+    expect(component.$('.dp-navigation .year').innerText).to.equal('2017')
+    expect(component.$('.dp-navigation .month').innerText).to.equal('Dec')
 
-    fireEvent($('su-datepicker button.ui.icon.button')[0], 'click')
-    $('su-datepicker .dp-day').length.should.equal(0)
-    $('su-datepicker .dp-month').length.should.equal(20)
-    $('su-datepicker .dp-month:first').text().trim().should.equal('2008')
+    fireEvent(component.$('button.ui.icon.button'), 'click')
+    expect(component.$('.dp-day')).to.be.undefined
+    expect(component.$$('.dp-month').length).to.equal(20)
+    expect(component.$('.dp-month').innerText.trim()).to.equal('2008')
 
-    fireEvent($('su-datepicker .dp-navigation .next')[0], 'click')
-    $('su-datepicker .dp-month:first').text().trim().should.equal('2028')
+    fireEvent(component.$('.dp-navigation .next'), 'click')
+    expect(component.$('.dp-month').innerText.trim()).to.equal('2028')
 
-    fireEvent($('su-datepicker .dp-navigation .prev')[0], 'click')
-    $('su-datepicker .dp-month:first').text().trim().should.equal('2008')
+    fireEvent(component.$('.dp-navigation .prev'), 'click')
+    expect(component.$('.dp-month').innerText.trim()).to.equal('2008')
 
-    fireEvent($('su-datepicker .dp-month:first .button')[0], 'click')
-    tag.currentDate.getTime().should.equal(new Date(2008, 11, 1).getTime())
+    fireEvent(component.$('.dp-month .button'), 'click')
+    expect(component.$('.dp-navigation .year').innerText).to.equal('2008')
+    expect(component.$('.dp-navigation .month').innerText).to.equal('Dec')
 
-    $('su-datepicker .dp-day').length.should.equal(0)
-    $('su-datepicker .dp-month').length.should.equal(12)
+    expect(component.$('.dp-day')).to.undefined
+    expect(component.$$('.dp-month').length).to.equal(12)
 
-    fireEvent($('su-datepicker .dp-month .ui.button:eq(0)')[0], 'click')
-    fireEvent($('su-datepicker .dp-day .ui.button:eq(5)')[0], 'click')
-    fireEvent($('su-datepicker button.ui.icon.button')[0], 'click')
-    $('su-datepicker .dp-month:first').text().trim().should.equal('1999')
+    fireEvent(component.$('.dp-month .ui.button'), 'click')
+    fireEvent(component.$$('.dp-day .ui.button')[5], 'click')
+    fireEvent(component.$('button.ui.icon.button'), 'click')
+    expect(component.$('.dp-month').innerText.trim()).to.equal('1999')
   })
 
   it('year range option', function () {
-    $('body').append('<su-datepicker />')
     mount({ currentDate: new Date(2017, 11, 1), startMode: "year", yearRange: "40" })
-    tag.currentDate.getTime().should.equal(new Date(2017, 11, 1).getTime())
+    expect(component.$('.dp-navigation .year').innerText).to.equal('2017')
+    expect(component.$('.dp-navigation .month').innerText).to.equal('Dec')
 
-    $('su-datepicker .dp-day').length.should.equal(0)
-    $('su-datepicker .dp-month').length.should.equal(40)
-    $('su-datepicker .dp-month:first').text().trim().should.equal('1998')
+    expect(component.$('.dp-day')).to.be.undefined
+    expect(component.$$('.dp-month').length).to.equal(40)
+    expect(component.$('.dp-month').innerText.trim()).to.equal('1998')
 
-    fireEvent($('su-datepicker .dp-navigation .next')[0], 'click')
-    $('su-datepicker .dp-month:first').text().trim().should.equal('2038')
+    fireEvent(component.$('.dp-navigation .next'), 'click')
+    expect(component.$('.dp-month').innerText.trim()).to.equal('2038')
 
-    fireEvent($('su-datepicker .dp-navigation .prev')[0], 'click')
-    $('su-datepicker .dp-month:first').text().trim().should.equal('1998')
+    fireEvent(component.$('.dp-navigation .prev'), 'click')
+    expect(component.$('.dp-month').innerText.trim()).to.equal('1998')
 
-    fireEvent($('su-datepicker .dp-month:first .button')[0], 'click')
-    tag.currentDate.getTime().should.equal(new Date(1998, 11, 1).getTime())
+    fireEvent(component.$('.dp-month .button'), 'click')
+    expect(component.$('.dp-navigation .year').innerText).to.equal('1998')
+    expect(component.$('.dp-navigation .month').innerText).to.equal('Dec')
 
-    $('su-datepicker .dp-day').length.should.equal(0)
-    $('su-datepicker .dp-month').length.should.equal(12)
+    expect(component.$('.dp-day')).to.be.undefined
+    expect(component.$$('.dp-month').length).to.equal(12)
   })
 
   it('update value', function () {
-    $('body').append('<su-datepicker />')
     mount({
       pattern: 'YYYY/MM/DD',
       value: new Date(2017, 11, 1)
     })
 
-    $('su-datepicker .dp-navigation .month').text().trim().should.equal('Dec')
-    $('su-datepicker .dp-navigation .year').text().trim().should.equal('2017')
-    tag.value.should.equal('2017/12/01')
+    expect(component.$('.dp-navigation .month').innerText.trim()).to.equal('Dec')
+    expect(component.$('.dp-navigation .year').innerText.trim()).to.equal('2017')
+    expect(component.$('su-datepicker').getAttribute('value')).to.equal('2017-12-01')
+    expect(component.$('su-datepicker').getAttribute('formated-value')).to.equal('2017/12/01')
 
-    tag.value = "2018/01/01"
-    tag.update()
-    $('su-datepicker .dp-navigation .month').text().trim().should.equal('Jan')
-    $('su-datepicker .dp-navigation .year').text().trim().should.equal('2018')
-    tag.value.should.equal('2018/01/01')
+    component.value = "2018/01/01"
+    component.update()
+    expect(component.$('.dp-navigation .month').innerText.trim()).to.equal('Jan')
+    expect(component.$('.dp-navigation .year').innerText.trim()).to.equal('2018')
+    expect(component.$('su-datepicker').getAttribute('value')).to.equal('2018-01-01')
+    expect(component.$('su-datepicker').getAttribute('formated-value')).to.equal('2018/01/01')
 
-    tag.value = new Date(2018, 0, 2)
-    tag.update()
-    $('su-datepicker .dp-navigation .month').text().trim().should.equal('Jan')
-    $('su-datepicker .dp-navigation .year').text().trim().should.equal('2018')
-    tag.value.should.equal('2018/01/02')
-
-    tag.valueAsDate = new Date(2018, 0, 3)
-    tag.update()
-    $('su-datepicker .dp-navigation .month').text().trim().should.equal('Jan')
-    $('su-datepicker .dp-navigation .year').text().trim().should.equal('2018')
-    tag.value.should.equal('2018/01/03')
+    component.value = new Date(2018, 0, 2)
+    component.update()
+    expect(component.$('.dp-navigation .month').innerText.trim()).to.equal('Jan')
+    expect(component.$('.dp-navigation .year').innerText.trim()).to.equal('2018')
+    expect(component.$('su-datepicker').getAttribute('value')).to.equal('2018-01-02')
+    expect(component.$('su-datepicker').getAttribute('formated-value')).to.equal('2018/01/02')
   })
 
-  it('update value same day', function () {
-    $('body').append('<su-datepicker />')
-    mount({
-      pattern: 'YYYY/MM/DD',
-      value: new Date(2018, 0, 1)
-    })
+  // it('update value same day', function () {
+  //   mount({
+  //     pattern: 'YYYY/MM/DD',
+  //     value: new Date(2018, 0, 1)
+  //   })
 
-    $('su-datepicker .dp-navigation .month').text().trim().should.equal('Jan')
-    $('su-datepicker .dp-navigation .year').text().trim().should.equal('2018')
-    tag.value.should.equal('2018/01/01')
+  //   expect(component.$('.dp-navigation .month').innerText.trim()).to.equal('Jan')
+  //   expect(component.$('.dp-navigation .year').innerText.trim()).to.equal('2018')
+  //   expect(component.$('su-datepicker').getAttribute('value')).to.equal('2018-01-01')
+  //   expect(component.$('su-datepicker').getAttribute('formated-value')).to.equal('2018/01/01')
 
-    tag.value = "2018/01/01"
-    tag.update()
-    $('su-datepicker .dp-navigation .month').text().trim().should.equal('Jan')
-    $('su-datepicker .dp-navigation .year').text().trim().should.equal('2018')
-    tag.value.should.equal('2018/01/01')
-
-    tag.value = new Date(2018, 0, 1)
-    tag.update()
-    $('su-datepicker .dp-navigation .month').text().trim().should.equal('Jan')
-    $('su-datepicker .dp-navigation .year').text().trim().should.equal('2018')
-    tag.value.should.equal('2018/01/01')
-
-    tag.valueAsDate = new Date(2018, 0, 1)
-    tag.update()
-    $('su-datepicker .dp-navigation .month').text().trim().should.equal('Jan')
-    $('su-datepicker .dp-navigation .year').text().trim().should.equal('2018')
-    tag.value.should.equal('2018/01/01')
-  })
+  //   component.value = "2018/01/01"
+  //   component.update()
+  //   expect(component.$('.dp-navigation .month').innerText.trim()).to.equal('Jan')
+  //   expect(component.$('.dp-navigation .year').innerText.trim()).to.equal('2018')
+  //   expect(component.$('su-datepicker').getAttribute('value')).to.equal('2018-01-01')
+  //   expect(component.$('su-datepicker').getAttribute('formated-value')).to.equal('2018/01/01')
+  // })
 
   it('update value with pattern', function () {
-    $('body').append('<su-datepicker />')
     mount({
       pattern: 'MMMM D, YYYY',
       value: new Date(2017, 11, 1)
     })
 
-    $('su-datepicker .dp-navigation .month').text().trim().should.equal('Dec')
-    $('su-datepicker .dp-navigation .year').text().trim().should.equal('2017')
-    tag.value.should.equal('December 1, 2017')
+    expect(component.$('.dp-navigation .month').innerText.trim()).to.equal('Dec')
+    expect(component.$('.dp-navigation .year').innerText.trim()).to.equal('2017')
+    expect(component.$('su-datepicker').getAttribute('value')).to.equal('2017-12-01')
+    expect(component.$('su-datepicker').getAttribute('formated-value')).to.equal('December 1, 2017')
 
-    tag.value = "January 1, 2018"
-    tag.update()
-    $('su-datepicker .dp-navigation .month').text().trim().should.equal('Jan')
-    $('su-datepicker .dp-navigation .year').text().trim().should.equal('2018')
-    tag.value.should.equal('January 1, 2018')
+    component.value = "January 1, 2018"
+    component.update()
+    expect(component.$('.dp-navigation .month').innerText.trim()).to.equal('Jan')
+    expect(component.$('.dp-navigation .year').innerText.trim()).to.equal('2018')
+    expect(component.$('su-datepicker').getAttribute('value')).to.equal('2018-01-01')
+    expect(component.$('su-datepicker').getAttribute('formated-value')).to.equal('January 1, 2018')
 
-    tag.value = new Date(2018, 0, 2)
-    tag.update()
-    $('su-datepicker .dp-navigation .month').text().trim().should.equal('Jan')
-    $('su-datepicker .dp-navigation .year').text().trim().should.equal('2018')
-    tag.value.should.equal('January 2, 2018')
-
-    tag.valueAsDate = new Date(2018, 0, 3)
-    tag.update()
-    $('su-datepicker .dp-navigation .month').text().trim().should.equal('Jan')
-    $('su-datepicker .dp-navigation .year').text().trim().should.equal('2018')
-    tag.value.should.equal('January 3, 2018')
-  })
-
-  it('update option value', function () {
-    $('body').append('<su-datepicker />')
-    mount({
-      value: new Date(2017, 11, 1)
-    })
-
-    $('su-datepicker .dp-navigation .month').text().trim().should.equal('Dec')
-    $('su-datepicker .dp-navigation .year').text().trim().should.equal('2017')
-
-    tag.opts.riotValue = new Date(2018, 0, 1)
-    tag.update()
-    $('su-datepicker .dp-navigation .month').text().trim().should.equal('Jan')
-    $('su-datepicker .dp-navigation .year').text().trim().should.equal('2018')
+    component.value = new Date(2018, 0, 2)
+    component.update()
+    expect(component.$('.dp-navigation .month').innerText.trim()).to.equal('Jan')
+    expect(component.$('.dp-navigation .year').innerText.trim()).to.equal('2018')
+    expect(component.$('su-datepicker').getAttribute('value')).to.equal('2018-01-02')
+    expect(component.$('su-datepicker').getAttribute('formated-value')).to.equal('January 2, 2018')
   })
 
   it('read-only option', function () {
-    $('body').append('<su-datepicker class="read-only" />')
-    mount()
-    fireEvent($('su-datepicker .dp-day .ui.button:first')[0], 'click')
-    spyOnClick.should.have.been.callCount(0)
-    spyOnChange.should.have.been.callCount(0)
+    mount({
+      class: 'read-only'
+    })
+    fireEvent(component.$('.dp-day .ui.button'), 'click')
+    expect(spyOnClick).to.have.been.not.called
+    expect(spyOnChange).to.have.been.not.called
   })
 
   it('popup datepicker read-only option', function () {
-    $('body').append('<su-datepicker class="read-only" />')
     mount({
+      class: 'read-only',
       popup: true
     })
-    $('su-datepicker .menu').css('visibility').should.equal('hidden')
+    expect(component.$('.menu').classList.contains('hidden')).to.be.ok
 
-    fireEvent($('su-datepicker button.ui.icon.button')[0], 'click')
-    spyOnOpen.should.have.been.callCount(0)
-    $('su-datepicker .menu').css('visibility').should.equal('hidden')
+    fireEvent(component.$('button.ui.icon.button'), 'click')
+    expect(spyOnOpen).to.have.been.not.called
+    expect(component.$('.menu').classList.contains('hidden')).to.be.ok
   })
 
   it('today and clear event', function () {
-    $('body').append('<su-datepicker />')
     mount({
       popup: true
     })
-    fireEvent($('su-datepicker button.ui.icon.button')[0], 'click')
-    spyOnOpen.should.have.been.calledOnce
+    fireEvent(component.$('button.ui.icon.button'), 'click')
+    expect(spyOnOpen).to.have.been.calledOnce
 
-    fireEvent($('su-datepicker .dp-today .button')[0], 'click')
+    fireEvent(component.$('.dp-today .button'), 'click')
     const today = new Date()
-    tag.value.should.equal(`${today.getFullYear()}-${addZero(today.getMonth() + 1)}-${addZero(today.getDate())}`)
-    tag.valueAsDate.getFullYear().should.equal(today.getFullYear())
-    tag.valueAsDate.getMonth().should.equal(today.getMonth())
-    tag.valueAsDate.getDate().should.equal(today.getDate())
-    spyOnClick.should.have.been.callCount(0)
-    spyOnChange.should.have.been.calledOnce
-    spyOnClose.should.have.been.calledOnce
+    expect(component.$('su-datepicker').getAttribute('value')).to.equal(`${today.getFullYear()}-${addZero(today.getMonth() + 1)}-${addZero(today.getDate())}`)
+    expect(component.$('su-datepicker').getAttribute('formated-value')).to.equal(`${today.getFullYear()}-${addZero(today.getMonth() + 1)}-${addZero(today.getDate())}`)
+    // tag.valueAsDate.getFullYear().to.equal(today.getFullYear())
+    // tag.valueAsDate.getMonth().to.equal(today.getMonth())
+    // tag.valueAsDate.getDate().to.equal(today.getDate())
+    expect(spyOnClick).to.have.been.not.called
+    expect(spyOnChange).to.have.been.calledOnce
+    expect(spyOnClose).to.have.been.calledOnce
 
-    fireEvent($('su-datepicker button.ui.icon.button')[0], 'click')
-    spyOnOpen.should.have.been.calledTwice
+    fireEvent(component.$('button.ui.icon.button'), 'click')
+    expect(spyOnOpen).to.have.been.calledTwice
 
-    fireEvent($('su-datepicker .dp-clear .button')[0], 'click')
-    expect(tag.value).to.be.null
-    spyOnClick.should.have.been.callCount(0)
-    spyOnChange.should.have.been.calledTwice
-    spyOnClose.should.have.been.calledTwice
+    fireEvent(component.$('.dp-clear .button'), 'click')
+    expect(component.$('su-datepicker').getAttribute('value')).to.be.null
+    expect(component.$('su-datepicker').getAttribute('formated-value')).to.be.null
+    expect(spyOnClick).to.have.been.not.called
+    expect(spyOnChange).to.have.been.calledTwice
+    expect(spyOnClose).to.have.been.calledTwice
   })
 
   it('upward', function () {
-    $('body').append('<su-datepicker />')
     mount({
       popup: true,
       direction: 'upward'
     })
 
-    $('su-datepicker').hasClass('upward').should.equal(false)
+    expect(component.$('su-datepicker').classList.contains('upward')).to.equal(false)
 
-    fireEvent($('su-datepicker button.ui.icon.button')[0], 'click')
-    $('su-datepicker .ui.dropdown').hasClass('upward').should.equal(true)
+    fireEvent(component.$('button.ui.icon.button'), 'click')
+    expect(component.$('.ui.dropdown').classList.contains('upward')).to.equal(true)
   })
 
   it('downward', function () {
-    $('body').append('<su-datepicker />')
     mount({
       popup: true,
       direction: 'downward'
     })
-    $('su-datepicker').hasClass('upward').should.equal(false)
+    expect(component.$('su-datepicker').classList.contains('upward')).to.equal(false)
 
-    fireEvent($('su-datepicker button.ui.icon.button')[0], 'click')
-    $('su-datepicker .ui.dropdown').hasClass('upward').should.equal(false)
+    fireEvent(component.$('button.ui.icon.button'), 'click')
+    expect(component.$('.ui.dropdown').classList.contains('upward')).to.equal(false)
   })
 
   it('reset value', function () {
-    $('body').append('<su-datepicker />')
     mount()
-    expect(tag.value).to.be.null
-    expect(tag.valueAsDate).to.be.undefined
-    expect(tag.defaultValue).to.be.undefined
+    expect(component.$('su-datepicker').getAttribute('value')).to.be.null
+    expect(component.$('su-datepicker').getAttribute('formated-value')).to.be.null
+    expect(component.$('su-datepicker').getAttribute('changed')).to.be.not.ok
 
-    fireEvent($('su-datepicker .dp-today .button')[0], 'click')
+    fireEvent(component.$('.dp-today .button'), 'click')
     const today = new Date()
-    tag.value.should.equal(`${today.getFullYear()}-${addZero(today.getMonth() + 1)}-${addZero(today.getDate())}`)
-    tag.valueAsDate.getFullYear().should.equal(today.getFullYear())
-    tag.valueAsDate.getMonth().should.equal(today.getMonth())
-    tag.valueAsDate.getDate().should.equal(today.getDate())
-    expect(tag.defaultValue).to.be.undefined
+    expect(component.$('su-datepicker').getAttribute('value')).to.equal(`${today.getFullYear()}-${addZero(today.getMonth() + 1)}-${addZero(today.getDate())}`)
+    expect(component.$('su-datepicker').getAttribute('formated-value')).to.equal(`${today.getFullYear()}-${addZero(today.getMonth() + 1)}-${addZero(today.getDate())}`)
+    expect(component.$('su-datepicker').getAttribute('changed')).to.be.ok
 
-    tag.reset()
-    expect(tag.value).to.be.null
-    expect(tag.valueAsDate).to.be.undefined
-    expect(tag.defaultValue).to.be.undefined
+    fireEvent(component.$('#reset'), 'click')
+    expect(component.$('su-datepicker').getAttribute('value')).to.be.null
+    expect(component.$('su-datepicker').getAttribute('formated-value')).to.be.null
+    expect(component.$('su-datepicker').getAttribute('changed')).to.be.not.ok
   })
 
   it('reset default value', function () {
-    $('body').append('<su-datepicker />')
     const date = new Date(2017, 0, 1)
     mount({ value: date })
 
-    tag.valueAsDate.getTime().should.equal(date.getTime())
-    tag.defaultValue.getTime().should.equal(date.getTime())
-    tag.changed().should.equal(false)
+    expect(component.$('.dp-navigation .month').innerText.trim()).to.equal('Jan')
+    expect(component.$('.dp-navigation .year').innerText.trim()).to.equal('2017')
+    expect(component.$('su-datepicker').getAttribute('value')).to.equal('2017-01-01')
+    expect(component.$('su-datepicker').getAttribute('formated-value')).to.equal('2017-01-01')
+    expect(component.$('su-datepicker').getAttribute('changed')).to.be.not.ok
 
-    fireEvent($('su-datepicker .dp-today .button')[0], 'click')
+    fireEvent(component.$('.dp-today .button'), 'click')
     const today = new Date()
-    tag.value.should.equal(`${today.getFullYear()}-${addZero(today.getMonth() + 1)}-${addZero(today.getDate())}`)
-    tag.valueAsDate.getFullYear().should.equal(today.getFullYear())
-    tag.valueAsDate.getMonth().should.equal(today.getMonth())
-    tag.valueAsDate.getDate().should.equal(today.getDate())
-    tag.defaultValue.getTime().should.equal(date.getTime())
-    tag.changed().should.equal(true)
+    expect(component.$('su-datepicker').getAttribute('value')).to.equal(`${today.getFullYear()}-${addZero(today.getMonth() + 1)}-${addZero(today.getDate())}`)
+    expect(component.$('su-datepicker').getAttribute('formated-value')).to.equal(`${today.getFullYear()}-${addZero(today.getMonth() + 1)}-${addZero(today.getDate())}`)
+    expect(component.$('su-datepicker').getAttribute('changed')).to.be.ok
 
-    tag.reset()
-    tag.valueAsDate.getTime().should.equal(date.getTime())
-    tag.defaultValue.getTime().should.equal(date.getTime())
-    tag.changed().should.equal(false)
+    fireEvent(component.$('#reset'), 'click')
+    expect(component.$('su-datepicker').getAttribute('value')).to.equal('2017-01-01')
+    expect(component.$('su-datepicker').getAttribute('formated-value')).to.equal('2017-01-01')
+    expect(component.$('su-datepicker').getAttribute('changed')).to.be.not.ok
   })
 })
