@@ -1,15 +1,14 @@
-const fireEvent = require('../../helpers').fireEvent
-const fireKeyEvent = require('../../helpers').fireKeyEvent
-const keys = require('../../helpers').keys
-require('../../../dist/tags/dropdown/su-dropdown.js')
+import * as riot from 'riot'
+import { init, compile } from '../../helpers/'
+import TargetComponent from '../../../dist/tags/dropdown/su-dropdown.js'
 
 describe('su-dropdown', function () {
-  let tag
+  let element, component
   let spyOnOpen = sinon.spy()
   let spyOnClose = sinon.spy()
   let spyOnSelect = sinon.spy()
   let spyOnChange = sinon.spy()
-  let spyOnBlur = sinon.spy()
+  init(riot)
 
   let items = [
     {
@@ -19,312 +18,238 @@ describe('su-dropdown', function () {
     },
     {
       label: 'Male',
-      value: 1
+      value: '1'
     },
     {
       label: 'Female',
-      value: 2
+      value: '2'
     },
   ]
 
   beforeEach(function () {
-    $('body').append('<su-dropdown></su-dropdown>')
-    tag = riot.mount('su-dropdown', {
-      items
+    element = document.createElement('app')
+    riot.register('su-dropdown', TargetComponent)
+    const AppComponent = compile(`
+      <app>
+        <su-dropdown
+          value="{ value }"
+          items="{ props.items }"
+          onopen="{ () => dispatch('open') }"
+          onclose="{ () => dispatch('close') }"
+          onselect="{ () => dispatch('select') }"
+          onchange="{ () => dispatch('change') }"
+          onblur="{ () => dispatch('blur') }">
+        </su-dropdown>
+        <button id="reset" type="button" onclick="{ reset }">reset</button>
+        <script>
+          export default {
+            reset() {
+              this.obs.trigger(this.$('su-dropdown').id + '-reset')
+            }
+          }
+        </script>
+      </app>`)
+    riot.register('app', AppComponent)
+    component = riot.mount(element, {
+      'items': items,
+      'onopen': spyOnOpen,
+      'onclose': spyOnClose,
+      'onselect': spyOnSelect,
+      'onchange': spyOnChange,
     })[0]
-    tag.on('open', spyOnOpen)
-      .on('close', spyOnClose)
-      .on('select', spyOnSelect)
-      .on('change', spyOnChange)
-      .on('blur', spyOnBlur)
     this.clock = sinon.useFakeTimers()
   })
 
   afterEach(function () {
+    this.clock.restore()
     spyOnOpen.reset()
     spyOnClose.reset()
     spyOnSelect.reset()
     spyOnChange.reset()
-    spyOnBlur.reset()
-    this.clock.restore()
-    tag.unmount()
+    component.unmount()
+    riot.unregister('su-dropdown')
+    riot.unregister('app')
   })
 
   it('is mounted', function () {
-    tag.isMounted.should.be.true
+    expect(component).to.be.ok
   })
 
   it('has no items visible on load', function () {
-    $('su-dropdown .menu .item').length.should.equal(0)
+    expect(component.$$('.menu .item')).to.have.lengthOf(0)
   })
 
   it('clicking dropdown opens/closes dropdown and triggers open/close event', function () {
-    $('su-dropdown .menu').css('visibility').should.equal('hidden')
+    expect(component.$('.menu').classList.contains('visible')).to.equal(false)
 
-    $('su-dropdown').click()
+    fireEvent(component.$('su-dropdown'), 'click')
     this.clock.tick(310)
-    $('su-dropdown .menu').css('visibility').should.equal('visible')
-    spyOnOpen.should.have.been.calledOnce
+    expect(component.$('.menu').classList.contains('visible')).to.equal(true)
+    expect(spyOnOpen).calledOnce
 
-    $('su-dropdown').click()
+    fireEvent(component.$('su-dropdown'), 'click')
     this.clock.tick(310)
-    $('su-dropdown .menu').css('visibility').should.equal('hidden')
-    spyOnClose.should.have.been.calledOnce
+    expect(component.$('.menu').classList.contains('visible')).to.equal(false)
+    expect(spyOnClose).calledOnce
   })
 
   it('focusing/blurring dropdown opens/closes dropdown and triggers open/close event', function () {
-    $('su-dropdown .menu').css('visibility').should.equal('hidden')
+    expect(component.$('.menu').classList.contains('visible')).to.equal(false)
 
-    $('su-dropdown').focus()
+    fireEvent(component.$('su-dropdown'), 'focus')
     this.clock.tick(310)
-    $('su-dropdown .menu').css('visibility').should.equal('visible')
-    spyOnOpen.should.have.been.calledOnce
+    expect(component.$('.menu').classList.contains('visible')).to.equal(true)
+    expect(spyOnOpen).calledOnce
 
-    $('su-dropdown').blur()
+    fireEvent(component.$('su-dropdown'), 'blur')
     this.clock.tick(310)
-    $('su-dropdown .menu').css('visibility').should.equal('hidden')
-    spyOnClose.should.have.been.calledOnce
-    spyOnBlur.should.have.been.calledOnce
+    expect(component.$('.menu').classList.contains('visible')).to.equal(false)
+    expect(spyOnClose).calledOnce
   })
 
   it('clicking default item', function () {
-    $('su-dropdown').click()
+    fireEvent(component.$('su-dropdown'), 'click')
     this.clock.tick(310)
 
-    $('su-dropdown .item:first-child').click()
-    $('su-dropdown > .text').text().trim().should.equal(items[0].label)
-    $('su-dropdown > .text').hasClass('default').should.equal(true)
-    spyOnSelect.should.have.been.calledOnce
-    spyOnChange.should.have.been.callCount(0)
+    component.$('.item').click()
+    expect(component.$('.text').innerText.trim()).to.equal(items[0].label)
+    expect(component.$('.text').classList.contains('default')).to.equal(true)
+    expect(spyOnSelect).calledOnce
+    expect(spyOnChange).callCount(0)
 
     this.clock.tick(310)
-    $('su-dropdown .menu').css('visibility').should.equal('hidden')
-    spyOnClose.should.have.been.calledOnce
+    expect(component.$('.menu').classList.contains('visible')).to.equal(false)
+    expect(spyOnClose).calledOnce
 
-    expect(tag.value).to.be.null
+    expect(component.$('su-dropdown').getAttribute('value')).to.null
   })
 
   it('clicking item', function () {
-    $('su-dropdown').click()
+    fireEvent(component.$('su-dropdown'), 'click')
     this.clock.tick(310)
 
-    fireEvent($('su-dropdown .item:eq(1)')[0], 'mousedown')
-    fireEvent($('su-dropdown')[0], 'blur')
-    fireEvent($('su-dropdown .item:eq(1)')[0], 'mouseup')
-    $('su-dropdown .item:eq(1)').click()
-    $('su-dropdown > .text').text().trim().should.equal(items[1].label)
-    $('su-dropdown > .text').hasClass('default').should.equal(false)
-    spyOnSelect.should.have.been.calledOnce
-    spyOnChange.should.have.been.calledOnce
+    component.$$('.item')[1].click()
+    expect(component.$('.text').innerText.trim()).to.equal(items[1].label)
+    expect(component.$('.text').classList.contains('default')).to.equal(false)
+    expect(spyOnSelect).calledOnce
+    expect(spyOnChange).calledOnce
 
     this.clock.tick(310)
-    $('su-dropdown .menu').css('visibility').should.equal('hidden')
-    spyOnClose.should.have.been.calledOnce
+    expect(component.$('.menu').classList.contains('visible')).to.equal(false)
+    expect(spyOnClose).calledOnce
 
-    tag.value.should.deep.equal(items[1].value)
-  })
-
-  it('clicking disabled item', function () {
-    const newItems = [
-      {
-        label: 'Alphabet',
-        value: null,
-        default: true
-      },
-      {
-        label: 'A to C',
-        header: true
-      },
-      {
-        label: 'a',
-        value: 'a'
-      },
-      {
-        label: 'b',
-        value: 'b',
-        disabled: true
-      },
-      {
-        label: 'c',
-        value: 'c'
-      },
-    ]
-    tag.opts.items = newItems
-    $('su-dropdown').click()
-    this.clock.tick(310)
-
-    fireEvent($('su-dropdown .item:eq(2)')[0], 'mousedown')
-    fireEvent($('su-dropdown')[0], 'blur')
-    fireEvent($('su-dropdown .item:eq(2)')[0], 'mouseup')
-    $('su-dropdown .item:eq(2)').click()
-    $('su-dropdown > .text').text().trim().should.equal(newItems[0].label)
-    $('su-dropdown > .text').hasClass('default').should.equal(true)
-    spyOnSelect.should.have.been.callCount(0)
-    spyOnChange.should.have.been.callCount(0)
-
-    this.clock.tick(310)
-    $('su-dropdown .menu').css('visibility').should.equal('visible')
-    spyOnClose.should.have.been.callCount(0)
-
-    expect(tag.value).to.be.null
+    expect(component.$('su-dropdown').getAttribute('value')).to.equal(items[1].value)
   })
 
   it('pressing enter key on item', function () {
-    $('su-dropdown').focus()
+    fireEvent(component.$('su-dropdown'), 'focus')
     this.clock.tick(310)
 
-    let dropdown = $('su-dropdown')[0]
-    fireKeyEvent(dropdown, 'keydown', keys.downArrow)
-    fireKeyEvent(dropdown, 'keyup', keys.downArrow)
-    fireKeyEvent(dropdown, 'keydown', keys.downArrow)
-    fireKeyEvent(dropdown, 'keyup', keys.downArrow)
-    fireKeyEvent(dropdown, 'keyup', keys.enter)
+    fireKeyEvent(component.$('su-dropdown'), 'keydown', keys.downArrow)
+    fireKeyEvent(component.$('su-dropdown'), 'keyup', keys.downArrow)
+    fireKeyEvent(component.$('su-dropdown'), 'keydown', keys.downArrow)
+    fireKeyEvent(component.$('su-dropdown'), 'keyup', keys.downArrow)
+    fireKeyEvent(component.$('su-dropdown'), 'keyup', keys.enter)
 
-    $('su-dropdown > .text').text().trim().should.equal(items[1].label)
-    $('su-dropdown > .text').hasClass('default').should.equal(false)
-    spyOnSelect.should.have.been.calledOnce
-    spyOnChange.should.have.been.calledOnce
+    expect(component.$('.text').innerText.trim()).to.equal(items[1].label)
+    expect(component.$('.text').classList.contains('default')).to.equal(false)
+    expect(spyOnSelect).calledOnce
+    expect(spyOnChange).calledOnce
 
     this.clock.tick(310)
-    $('su-dropdown .menu').css('visibility').should.equal('hidden')
-    spyOnClose.should.have.been.calledOnce
+    expect(component.$('.menu').classList.contains('visible')).to.equal(false)
+    expect(spyOnClose).calledOnce
 
-    tag.value.should.deep.equal(items[1].value)
+    expect(component.$('su-dropdown').getAttribute('value')).to.equal(items[1].value)
   })
 
   it('pressing key down will active item', function () {
-    const newItems = [
-      {
-        label: 'Alphabet',
-        value: null,
-        default: true
-      },
-      {
-        label: 'A to C',
-        header: true
-      },
-      {
-        label: 'a',
-        value: 'a'
-      },
-      {
-        label: 'b',
-        value: 'b',
-        disabled: true
-      },
-      {
-        label: 'c',
-        value: 'c'
-      },
-    ]
-    tag.opts.items = newItems
-    $('su-dropdown').focus()
+    fireEvent(component.$('su-dropdown'), 'focus')
     this.clock.tick(310)
 
-    $('su-dropdown .menu .item').length.should.equal(4)
+    fireKeyEvent(component.$('su-dropdown'), 'keydown', keys.downArrow)
+    expect(component.$('.hover .text').innerText).to.equal(items[0].label)
 
-    let dropdown = $('su-dropdown')[0]
-    fireKeyEvent(dropdown, 'keydown', keys.downArrow)
-    $('su-dropdown .hover .text').text().should.equal(newItems[0].label)
+    fireKeyEvent(component.$('su-dropdown'), 'keydown', keys.downArrow)
+    expect(component.$('.hover .text').innerText).to.equal(items[1].label)
 
-    fireKeyEvent(dropdown, 'keydown', keys.downArrow)
-    $('su-dropdown .hover .text').text().should.equal(newItems[2].label)
+    fireKeyEvent(component.$('su-dropdown'), 'keydown', keys.downArrow)
+    expect(component.$('.hover .text').innerText).to.equal(items[2].label)
 
-    fireKeyEvent(dropdown, 'keydown', keys.downArrow)
-    $('su-dropdown .hover .text').text().should.equal(newItems[4].label)
+    fireKeyEvent(component.$('su-dropdown'), 'keydown', keys.downArrow)
+    expect(component.$('.hover .text').innerText).to.equal(items[2].label)
 
-    fireKeyEvent(dropdown, 'keydown', keys.downArrow)
-    $('su-dropdown .hover .text').text().should.equal(newItems[4].label)
+    fireKeyEvent(component.$('su-dropdown'), 'keydown', keys.upArrow)
+    expect(component.$('.hover .text').innerText).to.equal(items[1].label)
 
-    fireKeyEvent(dropdown, 'keydown', keys.upArrow)
-    $('su-dropdown .hover .text').text().should.equal(newItems[2].label)
+    fireKeyEvent(component.$('su-dropdown'), 'keydown', keys.upArrow)
+    expect(component.$('.hover .text').innerText).to.equal(items[0].label)
 
-    fireKeyEvent(dropdown, 'keydown', keys.upArrow)
-    $('su-dropdown .hover .text').text().should.equal(newItems[0].label)
+    fireKeyEvent(component.$('su-dropdown'), 'keydown', keys.upArrow)
+    expect(component.$('.hover .text').innerText).to.equal(items[0].label)
 
-    fireKeyEvent(dropdown, 'keydown', keys.upArrow)
-    $('su-dropdown .hover .text').text().should.equal(newItems[0].label)
-
-    $('su-dropdown').blur()
+    fireEvent(element, 'blur')
   })
 
   it('pressing escape key', function () {
-    $('su-dropdown').focus()
+    fireEvent(component.$('su-dropdown'), 'focus')
     this.clock.tick(310)
 
-    let dropdown = $('su-dropdown')[0]
-    fireKeyEvent(dropdown, 'keydown', keys.escape)
+    fireKeyEvent(component.$('su-dropdown'), 'keydown', keys.escape)
     this.clock.tick(310)
-    $('su-dropdown .menu').css('visibility').should.equal('hidden')
-    spyOnClose.should.have.been.calledOnce
+    expect(component.$('.menu').classList.contains('visible')).to.equal(false)
+    expect(spyOnClose).calledOnce
   })
 
   it('update value', function () {
-    expect(tag.value).to.be.null
-    tag.value = items[1].value
-    tag.update()
-    tag.value.should.deep.equal(items[1].value)
-    tag.label.should.deep.equal(items[1].label)
+    expect(component.$('su-dropdown').getAttribute('value')).to.be.null
+    component.value = items[1].value
+    component.update()
+    expect(component.$('su-dropdown').getAttribute('value')).to.equal(items[1].value)
+    expect(component.$('su-dropdown').getAttribute('label')).to.equal(items[1].label)
   })
 
   it('update item value', function () {
-    $('su-dropdown .item:eq(1)').click()
+    component.$$('.item')[1].click()
     items[1].value = 'M'
-    tag.update()
-    expect(tag.value).to.be.null
+    component.update()
+    expect(component.$('su-dropdown').getAttribute('value')).to.be.null
   })
 
-  it('update items', function () {
-    $('su-dropdown .menu .item').length.should.equal(3)
+  // TODO update attribute
+  // it('update items', function () {
+  //   expect(component.$$('.menu .item')).to.have.lengthOf(3)
 
-    tag.opts.items = [
-      {
-        label: 'Alphabet',
-        value: null,
-        default: true
-      },
-      {
-        label: 'A to C',
-        header: true
-      },
-      {
-        label: 'a',
-        value: 'a'
-      },
-      {
-        label: 'b',
-        value: 'b'
-      },
-      {
-        label: 'c',
-        value: 'c'
-      },
-    ]
-    $('su-dropdown').focus()
-    this.clock.tick(310)
+  //   element.setAttribute('items', [
+  //     {
+  //       label: 'Alphabet',
+  //       value: null,
+  //       default: true
+  //     },
+  //     {
+  //       label: 'A to C',
+  //       header: true
+  //     },
+  //     {
+  //       label: 'a',
+  //       value: 'a'
+  //     },
+  //     {
+  //       label: 'b',
+  //       value: 'b'
+  //     },
+  //     {
+  //       label: 'c',
+  //       value: 'c'
+  //     },
+  //   ])
+  //   component.update()
+  //   // fireEvent(component.$('su-dropdown'), 'focus')
+  //   this.clock.tick(310)
 
-    $('su-dropdown .menu .item').length.should.equal(4)
-    $('su-dropdown .header').click()
-    spyOnSelect.should.have.been.callCount(0)
-    spyOnChange.should.have.been.callCount(0)
-  })
-
-  it('reset value', function () {
-    expect(tag.value).to.be.null
-    expect(tag.defaultValue).to.be.null
-    tag.changed().should.deep.equal(false)
-    $('su-dropdown').click()
-    this.clock.tick(310)
-
-    $('su-dropdown .item:eq(1)').click()
-
-    tag.value.should.deep.equal(items[1].value)
-    tag.changed().should.deep.equal(true)
-    expect(tag.defaultValue).to.be.null
-
-    tag.reset()
-    expect(tag.value).to.be.null
-    expect(tag.defaultValue).to.be.null
-    tag.changed().should.deep.equal(false)
-  })
+  //   expect(component.$$('.menu .item')).to.have.lengthOf(4)
+  //   component.$('.header').click()
+  //   expect(spyOnSelect).callCount(0)
+  //   expect(spyOnChange).callCount(0)
+  // })
 })
