@@ -1,110 +1,126 @@
-const fireEvent = require('../../helpers').fireEvent
-require('../../../dist/tags/rating/su-rating.js')
+import * as riot from 'riot'
+import { init, compile } from '../../helpers/'
+import TargetComponent from '../../../dist/tags/rating/su-rating.js'
 
 describe('su-rating', function () {
-  let tag
+  let element, component
   const spyOnClick = sinon.spy()
   const spyOnChange = sinon.spy()
+  init(riot)
 
   const mount = opts => {
-    tag = riot.mount('su-rating', opts)[0]
-    tag.on('click', spyOnClick)
-    tag.on('change', spyOnChange)
+    const option = Object.assign({
+      'onclick': spyOnClick,
+      'onchange': spyOnChange,
+    }, opts)
+    element = document.createElement('app')
+    riot.register('su-rating', TargetComponent)
+    const AppComponent = compile(`
+      <app>
+        <su-rating
+          onclick="{ () => dispatch('click') }"
+          onchange="{ () => dispatch('change') }"
+        />
+        <button id="reset" type="button" onclick="{ reset }">reset</button>
+        <script>
+          export default {
+            reset() {
+              this.obs.trigger(this.$('su-rating').id + '-reset')
+            }
+          }
+        </script>
+      </app>`)
+    riot.register('app', AppComponent)
+    component = riot.mount(element, option)[0]
   }
-
-
-  beforeEach(function () {
-    $('body').append('<su-rating></su-rating>')
-  })
 
   afterEach(function () {
     spyOnClick.reset()
     spyOnChange.reset()
-    tag.unmount()
+    riot.unregister('su-rating')
+    riot.unregister('app')
   })
 
   it('is mounted', function () {
     mount()
-    tag.isMounted.should.be.true
+    expect(component).to.be.ok
   })
 
   it('click event', function () {
     mount()
-    $('su-rating i').length.should.equal(5)
-    tag.value.should.equal(0)
-    $('su-rating i:eq(0)').hasClass('active').should.equal(false)
-    $('su-rating i:eq(1)').hasClass('active').should.equal(false)
-    $('su-rating i:eq(2)').hasClass('active').should.equal(false)
-    $('su-rating i:eq(3)').hasClass('active').should.equal(false)
-    $('su-rating i:eq(4)').hasClass('active').should.equal(false)
-    spyOnClick.should.have.been.callCount(0)
-    spyOnChange.should.have.been.callCount(0)
+    expect(component.$$('su-rating i').length).to.equal(5)
+    expect(component.$('su-rating').getAttribute('value')).to.equal('0')
+    expect(component.$$('su-rating i')[0].classList.contains('active')).to.equal(false)
+    expect(component.$$('su-rating i')[1].classList.contains('active')).to.equal(false)
+    expect(component.$$('su-rating i')[2].classList.contains('active')).to.equal(false)
+    expect(component.$$('su-rating i')[3].classList.contains('active')).to.equal(false)
+    expect(component.$$('su-rating i')[4].classList.contains('active')).to.equal(false)
+    expect(spyOnClick).to.have.been.not.called
+    expect(spyOnChange).to.have.been.not.called
 
-    fireEvent($('su-rating i:eq(2)')[0], 'click')
-    tag.value.should.equal(3)
-    $('su-rating i:eq(0)').hasClass('active').should.equal(true)
-    $('su-rating i:eq(1)').hasClass('active').should.equal(true)
-    $('su-rating i:eq(2)').hasClass('active').should.equal(true)
-    $('su-rating i:eq(3)').hasClass('active').should.equal(false)
-    $('su-rating i:eq(4)').hasClass('active').should.equal(false)
-    spyOnClick.should.have.been.calledOnce
-    spyOnChange.should.have.been.calledOnce
+    fireEvent(component.$$('su-rating i')[2], 'click')
+    expect(component.$('su-rating').getAttribute('value')).to.equal('3')
+    expect(component.$$('su-rating i')[0].classList.contains('active')).to.equal(true)
+    expect(component.$$('su-rating i')[1].classList.contains('active')).to.equal(true)
+    expect(component.$$('su-rating i')[2].classList.contains('active')).to.equal(true)
+    expect(component.$$('su-rating i')[3].classList.contains('active')).to.equal(false)
+    expect(component.$$('su-rating i')[4].classList.contains('active')).to.equal(false)
+    expect(spyOnClick).to.have.been.calledOnce
+    expect(spyOnChange).to.have.been.calledOnce
 
-    fireEvent($('su-rating i:eq(2)')[0], 'click')
-    spyOnClick.should.have.been.calledTwice
-    spyOnChange.should.have.been.calledOnce
+    fireEvent(component.$$('su-rating i')[2], 'click')
+    expect(spyOnClick).to.have.been.calledTwice
+    expect(spyOnChange).to.have.been.calledOnce
 
-    fireEvent($('su-rating i:eq(0)')[0], 'click')
-    tag.value.should.equal(1)
-    $('su-rating i:eq(0)').hasClass('active').should.equal(true)
-    $('su-rating i:eq(1)').hasClass('active').should.equal(false)
-    $('su-rating i:eq(2)').hasClass('active').should.equal(false)
-    $('su-rating i:eq(3)').hasClass('active').should.equal(false)
-    $('su-rating i:eq(4)').hasClass('active').should.equal(false)
+    fireEvent(component.$$('su-rating i')[0], 'click')
+    expect(component.$('su-rating').getAttribute('value')).to.equal('1')
+    expect(component.$$('su-rating i')[0].classList.contains('active')).to.equal(true)
+    expect(component.$$('su-rating i')[1].classList.contains('active')).to.equal(false)
+    expect(component.$$('su-rating i')[2].classList.contains('active')).to.equal(false)
+    expect(component.$$('su-rating i')[3].classList.contains('active')).to.equal(false)
+    expect(component.$$('su-rating i')[4].classList.contains('active')).to.equal(false)
   })
 
   it('mouseover and mouseout', function () {
     mount()
-    $('su-rating i').length.should.equal(5)
-    tag.value.should.equal(0)
-    $('su-rating i:eq(0)').hasClass('selected').should.equal(false)
-    $('su-rating i:eq(1)').hasClass('selected').should.equal(false)
-    $('su-rating i:eq(2)').hasClass('selected').should.equal(false)
-    $('su-rating i:eq(3)').hasClass('selected').should.equal(false)
-    $('su-rating i:eq(4)').hasClass('selected').should.equal(false)
+    expect(component.$$('su-rating i').length).to.equal(5)
+    expect(component.$('su-rating').getAttribute('value')).to.equal('0')
+    expect(component.$$('su-rating i')[0].classList.contains('selected')).to.equal(false)
+    expect(component.$$('su-rating i')[1].classList.contains('selected')).to.equal(false)
+    expect(component.$$('su-rating i')[2].classList.contains('selected')).to.equal(false)
+    expect(component.$$('su-rating i')[3].classList.contains('selected')).to.equal(false)
+    expect(component.$$('su-rating i')[4].classList.contains('selected')).to.equal(false)
 
-    fireEvent($('su-rating i:eq(2)')[0], 'mouseover')
-    tag.value.should.equal(0)
-    $('su-rating i:eq(0)').hasClass('selected').should.equal(true)
-    $('su-rating i:eq(1)').hasClass('selected').should.equal(true)
-    $('su-rating i:eq(2)').hasClass('selected').should.equal(true)
-    $('su-rating i:eq(3)').hasClass('selected').should.equal(false)
-    $('su-rating i:eq(4)').hasClass('selected').should.equal(false)
+    fireEvent(component.$$('su-rating i')[2], 'mouseover')
+    expect(component.$('su-rating').getAttribute('value')).to.equal('0')
+    expect(component.$$('su-rating i')[0].classList.contains('selected')).to.equal(true)
+    expect(component.$$('su-rating i')[1].classList.contains('selected')).to.equal(true)
+    expect(component.$$('su-rating i')[2].classList.contains('selected')).to.equal(true)
+    expect(component.$$('su-rating i')[3].classList.contains('selected')).to.equal(false)
+    expect(component.$$('su-rating i')[4].classList.contains('selected')).to.equal(false)
 
-    fireEvent($('su-rating i:eq(2)')[0], 'mouseout')
-    tag.value.should.equal(0)
-    $('su-rating i:eq(0)').hasClass('selected').should.equal(false)
-    $('su-rating i:eq(1)').hasClass('selected').should.equal(false)
-    $('su-rating i:eq(2)').hasClass('selected').should.equal(false)
-    $('su-rating i:eq(3)').hasClass('selected').should.equal(false)
-    $('su-rating i:eq(4)').hasClass('selected').should.equal(false)
+    fireEvent(component.$$('su-rating i')[2], 'mouseout')
+    expect(component.$('su-rating').getAttribute('value')).to.equal('0')
+    expect(component.$$('su-rating i')[0].classList.contains('selected')).to.equal(false)
+    expect(component.$$('su-rating i')[1].classList.contains('selected')).to.equal(false)
+    expect(component.$$('su-rating i')[2].classList.contains('selected')).to.equal(false)
+    expect(component.$$('su-rating i')[3].classList.contains('selected')).to.equal(false)
+    expect(component.$$('su-rating i')[4].classList.contains('selected')).to.equal(false)
   })
 
   it('reset value', function () {
     mount()
 
-    tag.value.should.equal(0)
-    tag.defaultValue.should.equal(0)
-    tag.changed().should.deep.equal(false)
+    expect(component.$('su-rating').getAttribute('value')).to.equal('0')
+    expect(component.$('su-rating').getAttribute('changed')).to.be.not.ok
 
-    fireEvent($('su-rating i:eq(2)')[0], 'click')
-    tag.value.should.equal(3)
-    tag.defaultValue.should.equal(0)
-    tag.changed().should.deep.equal(true)
+    fireEvent(component.$$('su-rating i')[2], 'click')
+    component.update()
+    expect(component.$('su-rating').getAttribute('value')).to.equal('3')
+    expect(component.$('su-rating').getAttribute('changed')).to.be.ok
 
-    tag.reset()
-    tag.value.should.equal(0)
-    tag.defaultValue.should.equal(0)
-    tag.changed().should.deep.equal(false)
+    fireEvent(component.$('#reset'), 'click')
+    expect(component.$('su-rating').getAttribute('value')).to.equal('0')
+    expect(component.$('su-rating').getAttribute('changed')).to.be.not.ok
   })
 })
