@@ -1,56 +1,60 @@
-require('../../../dist/tags/checkbox/su-checkbox-group.js')
-require('../../../dist/tags/checkbox/su-checkbox.js')
+import * as riot from 'riot'
+import { init, compile } from '../../helpers/'
+import CheckboxGroupComponent from '../../../dist/tags/checkbox/su-checkbox-group.js'
+import CheckboxComponent from '../../../dist/tags/checkbox/su-checkbox.js'
 
 describe('su-checkbox-group-single', function () {
-  let tag
+  let element, component
   let spyOnChange = sinon.spy()
+  init(riot)
 
   beforeEach(function () {
-    const group = $('<su-checkbox-group></su-checkbox-group>')
-    group.append('<su-checkbox value="1">Checkbox choice1</su-checkbox>')
-    $('body').append(group)
-    tag = riot.mount('su-checkbox-group')[0]
-    tag.on('change', spyOnChange)
+    element = document.createElement('app')
+    riot.register('su-checkbox-group', CheckboxGroupComponent)
+    riot.register('su-checkbox', CheckboxComponent)
+    const AppComponent = compile(`
+      <app>
+        <su-checkbox-group
+          value="{ value }"
+          onchange="{ () => dispatch('change') }">
+          <su-checkbox value="1" />
+        </su-checkbox-group>
+      </app>`)
+    riot.register('app', AppComponent)
+    component = riot.mount(element, {
+      'onchange': spyOnChange
+    })[0]
   })
 
   afterEach(function () {
     spyOnChange.reset()
-    tag.unmount()
+    component.unmount()
+    riot.unregister('su-checkbox')
+    riot.unregister('su-checkbox-group')
+    riot.unregister('app')
   })
 
   it('is mounted', function () {
-    tag.isMounted.should.be.true
+    expect(component).to.be.ok
   })
 
   it('update value', function () {
-    tag.tags['su-checkbox'].checked.should.equal(false)
+    expect(component.$('su-checkbox').getAttribute("checked")).to.be.not.ok
 
-    tag.value = '1'
-    tag.update()
-    tag.tags['su-checkbox'].checked.should.equal(true)
+    component.value = "1"
+    component.update()
+    expect(component.$('su-checkbox').getAttribute("checked")).to.be.ok
 
-    tag.value = '2'
-    tag.update()
-    tag.tags['su-checkbox'].checked.should.equal(false)
-  })
-
-  it('update option', function () {
-    tag.tags['su-checkbox'].checked.should.equal(false)
-
-    tag.opts.riotValue = '1'
-    tag.update()
-    tag.tags['su-checkbox'].checked.should.equal(true)
-
-    tag.opts.riotValue = '2'
-    tag.update()
-    tag.tags['su-checkbox'].checked.should.equal(false)
+    component.value = "2"
+    component.update()
+    expect(component.$('su-checkbox').getAttribute("checked")).to.be.not.ok
   })
 
   it('click checkbox', function () {
-    tag.tags['su-checkbox'].checked.should.equal(false)
+    expect(component.$('su-checkbox').getAttribute("checked")).to.be.not.ok
 
-    $('su-checkbox:eq(0) input').click()
-    tag.tags['su-checkbox'].checked.should.equal(true)
-    spyOnChange.should.have.been.calledOnce
+    component.$('su-checkbox input').click()
+    expect(component.$('su-checkbox').getAttribute("checked")).to.be.ok
+    expect(spyOnChange).to.have.been.calledOnce
   })
 })
