@@ -1,69 +1,84 @@
-require('../../../dist/tags/radio/su-radio.js')
-require('../../../dist/tags/radio/su-radio-group.js')
+import * as riot from 'riot'
+import { init, compile } from '../../helpers/'
+import RadioGroupComponent from '../../../dist/tags/radio/su-radio-group.js'
+import RadioComponent from '../../../dist/tags/radio/su-radio.js'
 
-describe('su-radio-group-update', function () {
-  let tag
+describe('su-radio-group', function () {
+  let element, component
   let spyOnChange = sinon.spy()
+  init(riot)
 
   beforeEach(function () {
-    const group = $('<su-radio-group></su-radio-group>')
-    group.append('<su-radio each="{ item in items }" value="{ item }" label="Radio choice{ item }"></su-radio>')
-    $('body').append(group)
-    tag = riot.mount('su-radio-group')[0]
-    tag.on('change', spyOnChange)
+    element = document.createElement('app')
+    riot.register('su-radio-group', RadioGroupComponent)
+    riot.register('su-radio', RadioComponent)
+    const AppComponent = compile(`
+      <app>
+        <su-radio-group
+          value="{ value }"
+          onchange="{ () => dispatch('change') }">
+          <su-radio each="{ item in items }" value="{ item }" label="Radio choice{ item }"></su-radio>
+        </su-radio-group>
+      </app>`)
+    riot.register('app', AppComponent)
+    component = riot.mount(element, {
+      'onchange': spyOnChange
+    })[0]
   })
 
   afterEach(function () {
     spyOnChange.reset()
-    tag.unmount()
+    component.unmount()
+    riot.unregister('su-radio')
+    riot.unregister('su-radio-group')
+    riot.unregister('app')
   })
 
   it('is mounted', function () {
-    tag.isMounted.should.be.true
+    expect(component).to.be.ok
   })
 
   it('update value', function () {
-    expect(tag.tags['su-radio']).to.be.undefined
-    tag.update({ items: [1, 2] })
-    tag.update()
+    expect(component.$$('su-radio').length).to.equal(0)
+    component.items = [1, 2]
+    component.update()
 
-    tag.tags['su-radio'].length.should.equal(2)
-    tag.tags['su-radio'][0].checked.should.equal(false)
-    tag.tags['su-radio'][1].checked.should.equal(false)
+    expect(component.$$('su-radio')[0].hasAttribute("checked")).to.be.not.ok
+    expect(component.$$('su-radio')[1].hasAttribute("checked")).to.be.not.ok
 
-    tag.value = '1'
-    tag.update()
-    tag.tags['su-radio'][0].checked.should.equal(true)
-    tag.tags['su-radio'][1].checked.should.equal(false)
-    spyOnChange.should.have.been.calledOnce
+    component.value = "1"
+    component.update()
+    expect(component.$$('su-radio')[0].getAttribute("checked")).to.be.ok
+    expect(component.$$('su-radio')[1].hasAttribute("checked")).to.be.not.ok
+    expect(spyOnChange).to.have.been.calledOnce
 
-    tag.value = '2'
-    tag.update()
-    tag.tags['su-radio'][0].checked.should.equal(false)
-    tag.tags['su-radio'][1].checked.should.equal(true)
-    spyOnChange.should.have.been.calledTwice
+    component.value = "2"
+    component.update()
+    expect(component.$$('su-radio')[0].hasAttribute("checked")).to.be.not.ok
+    expect(component.$$('su-radio')[1].getAttribute("checked")).to.be.ok
+    expect(spyOnChange).to.have.been.calledTwice
 
-    tag.update({ items: null })
-    expect(tag.tags['su-radio']).to.be.undefined
+    component.items = null
+    component.update()
+    expect(component.$$('su-radio').length).to.equal(0)
   })
 
-  it('click checkbox', function () {
-    expect(tag.tags['su-radio']).to.be.undefined
-    tag.update({ items: [1, 2] })
-    tag.update()
+  it('click radio', function () {
+    expect(component.$$('su-radio').length).to.equal(0)
+    component.items = [1, 2]
+    component.update()
 
-    tag.tags['su-radio'].length.should.equal(2)
-    tag.tags['su-radio'][0].checked.should.equal(false)
-    tag.tags['su-radio'][1].checked.should.equal(false)
+    expect(component.$$('su-radio')[0].hasAttribute("checked")).to.be.not.ok
+    expect(component.$$('su-radio')[1].hasAttribute("checked")).to.be.not.ok
 
-    $('su-radio:eq(0) input').click()
-    tag.tags['su-radio'][0].checked.should.equal(true)
-    tag.tags['su-radio'][1].checked.should.equal(false)
-    spyOnChange.should.have.been.calledOnce
+    component.$$('su-radio input')[0].click()
+    expect(component.$$('su-radio')[0].getAttribute("checked")).to.be.ok
+    expect(component.$$('su-radio')[1].hasAttribute("checked")).to.be.not.ok
+    expect(spyOnChange).to.have.been.calledOnce
 
-    $('su-radio:eq(1) input').click()
-    tag.tags['su-radio'][0].checked.should.equal(false)
-    tag.tags['su-radio'][1].checked.should.equal(true)
-    spyOnChange.should.have.been.calledTwice
+    component.$$('su-radio input')[1].click()
+    expect(component.$$('su-radio')[0].hasAttribute("checked")).to.be.not.ok
+    expect(component.$$('su-radio')[1].getAttribute("checked")).to.be.ok
+    expect(spyOnChange).to.have.been.calledTwice
   })
 })
