@@ -1,77 +1,77 @@
-require('../../../dist/tags/table/su-table.js')
-require('../../../dist/tags/table/su-th.js')
+import * as riot from 'riot'
+import { init, compile } from '../../helpers/'
+import TableComponent from '../../../dist/tags/table/su-table.js'
+import ThComponent from '../../../dist/tags/table/su-th.js'
 
 describe('su-table-option', function () {
-  let tag
+  let element, component
+  init(riot)
+
+  beforeEach(function () {
+    element = document.createElement('app')
+    riot.register('su-table', TableComponent)
+    riot.register('su-th', ThComponent)
+  })
 
   const mount = (html, opts) => {
-    $('body').append(html)
-
-    tag = riot.mount('su-table', opts)[0]
+    const AppComponent = compile(html)
+    riot.register('app', AppComponent)
+    component = riot.mount(element, opts)[0]
   }
 
   afterEach(function () {
-    tag.unmount()
+    riot.unregister('su-th')
+    riot.unregister('su-table')
+    riot.unregister('app')
   })
 
   it('is mounted', function () {
-    const html = $('<su-table></su-table>')
-    html
-      .append('<su-th field="name">Name</su-th>')
-    mount(html)
-    tag.isMounted.should.be.true
+    mount('<app><su-table><su-th field="name">Name</su-th></su-table></app>')
+    expect(component).to.be.ok
   })
 
   it('single-header', function () {
-    const html = $('<su-table></su-table>')
-    html
-      .append('<su-th field="name">Name</su-th>')
-    mount(html, {
-      data: [
-        { name: 'John', age: 15, gender: 'Male' },
-        { name: 'Amber', age: 40, gender: 'Female' },
-        { name: 'Terry', age: 25, gender: 'Male' },
-        { name: 'Leslie', age: 25 },
-        { name: 'Ben', age: 70, gender: 'Male' },
-      ]
-    })
+    mount(
+      `<app>
+        <su-table data="{ props.data }">
+          <su-th field="name">Name</su-th>
+        </su-table>
+      </app>`,
+      {
+        data: [
+          { name: 'John', age: 15, gender: 'Male' },
+          { name: 'Amber', age: 40, gender: 'Female' },
+          { name: 'Terry', age: 25, gender: 'Male' },
+          { name: 'Leslie', age: 25 },
+          { name: 'Ben', age: 70, gender: 'Male' },
+        ]
+      })
 
     // name
-    $('su-table su-th:first').click()
-    tag.opts.data[0].name.should.equal('Amber')
-    tag.opts.data[0]['su-table-index'].should.equal(1)
-    tag.opts.data[1]['su-table-index'].should.equal(4)
-    tag.opts.data[2]['su-table-index'].should.equal(0)
-    tag.opts.data[3]['su-table-index'].should.equal(3)
-    tag.opts.data[4]['su-table-index'].should.equal(2)
-    $('su-table su-th:first').hasClass('sorted').should.equal(true)
-    $('su-table su-th:first').hasClass('ascending').should.equal(true)
-    $('su-table su-th:first').hasClass('descending').should.equal(false)
+    fireEvent(component.$$('su-th')[0], 'click')
+    expect(component.props.data[0].name).to.equal('Amber')
+    expect(component.props.data[0]['su-table-index']).to.equal(1)
+    expect(component.$('su-table su-th').classList.contains('sorted')).to.equal(true)
+    expect(component.$('su-table su-th').classList.contains('ascending')).to.equal(true)
+    expect(component.$('su-table su-th').classList.contains('descending')).to.equal(false)
 
-    // name desc
-    $('su-table su-th:first').click()
-    tag.opts.data[0].name.should.equal('Terry')
-    $('su-table su-th:first').hasClass('sorted').should.equal(true)
-    $('su-table su-th:eq(1)').hasClass('sorted').should.equal(false)
-    $('su-table su-th:eq(2)').hasClass('sorted').should.equal(false)
-    $('su-table su-th:eq(3)').hasClass('sorted').should.equal(false)
-    $('su-table su-th:first').hasClass('ascending').should.equal(false)
-    $('su-table su-th:first').hasClass('descending').should.equal(true)
+    // // name desc
+    fireEvent(component.$$('su-th')[0], 'click')
+    expect(component.props.data[0].name).to.equal('Terry')
+    expect(component.$('su-table su-th').classList.contains('sorted')).to.equal(true)
+    expect(component.$('su-table su-th').classList.contains('ascending')).to.equal(false)
+    expect(component.$('su-table su-th').classList.contains('descending')).to.equal(true)
 
-    // default
-    $('su-table su-th:first').click()
-    tag.opts.data[0].name.should.equal('John')
-    $('su-table su-th:first').hasClass('sorted').should.equal(false)
-    $('su-table su-th:eq(1)').hasClass('sorted').should.equal(false)
-    $('su-table su-th:eq(2)').hasClass('sorted').should.equal(false)
-    $('su-table su-th:eq(3)').hasClass('sorted').should.equal(false)
-    $('su-table su-th:first').hasClass('ascending').should.equal(false)
-    $('su-table su-th:first').hasClass('descending').should.equal(false)
+    // // default
+    fireEvent(component.$$('su-th')[0], 'click')
+    expect(component.props.data[0].name).to.equal('John')
+    expect(component.$('su-table su-th').classList.contains('sorted')).to.equal(false)
+    expect(component.$('su-table su-th').classList.contains('ascending')).to.equal(false)
+    expect(component.$('su-table su-th').classList.contains('descending')).to.equal(false)
   })
 
   it('none-header', function () {
-    const html = $('<su-table></su-table>')
-    mount(html, {
+    mount('<app><su-table data="{ props.data }"></su-table></app>', {
       data: [
         { name: 'John', age: 15, gender: 'Male' },
         { name: 'Amber', age: 40, gender: 'Female' },
@@ -81,12 +81,6 @@ describe('su-table-option', function () {
       ]
     })
 
-    tag.opts.data[0].name.should.equal('John')
-    $('su-table su-th:first').hasClass('sorted').should.equal(false)
-    $('su-table su-th:eq(1)').hasClass('sorted').should.equal(false)
-    $('su-table su-th:eq(2)').hasClass('sorted').should.equal(false)
-    $('su-table su-th:eq(3)').hasClass('sorted').should.equal(false)
-    $('su-table su-th:first').hasClass('ascending').should.equal(false)
-    $('su-table su-th:first').hasClass('descending').should.equal(false)
+    expect(component.props.data[0].name).to.equal('John')
   })
 })
