@@ -1,95 +1,99 @@
-require('../../../dist/tags/toast/su-toast.js')
-require('../../../dist/tags/toast/su-toast-item.js')
-
-riot.mixin('semantic-ui', {})
+import * as riot from 'riot'
+import { init, compile } from '../../helpers/'
+import TargetComponent from '../../../dist/tags/toast/su-toast.js'
+import TargetItemComponent from '../../../dist/tags/toast/su-toast-item.js'
 
 describe('su-toast', function () {
-  let tag
-  let app
-  let mount = () => {
-    tag = riot.mount('su-toast')[0]
-    app = riot.mount('app')[0]
+  let element, component
+  init(riot)
+
+  const mount = () => {
+    element = document.createElement('app')
+    riot.register('su-toast', TargetComponent)
+    riot.register('su-toast-item', TargetItemComponent)
+    const AppComponent = compile(`
+      <app>
+        <su-toast>
+        </su-toast>
+      </app>`)
+    riot.register('app', AppComponent)
+    component = riot.mount(element)[0]
   }
 
   beforeEach(function () {
     this.clock = sinon.useFakeTimers()
-    riot.mixin('semantic-ui', { observable: riot.observable() })
-    riot.tag('app')
-    $('body').append(`
-      <su-toast></su-toast>
-      <app></app>
-    `)
   })
 
   afterEach(function () {
+    riot.unregister('su-toast-item')
+    riot.unregister('su-toast')
+    riot.unregister('app')
     this.clock.restore()
-    tag.unmount()
-    app.unmount()
   })
 
   it('is mounted', function () {
     mount()
-    tag.isMounted.should.be.true
+    expect(component).to.be.ok
   })
 
   it('opens/auto closes toast', function () {
     mount()
-    $('su-toast .item').length.should.equal(0)
+    expect(component.$('su-toast .item')).to.be.undefined
 
-    app.suToast('hello')
-    tag.items[0].messages[0].should.equal('hello')
-    $('su-toast .item').length.should.equal(1)
+    component.suToast('hello')
+    expect(component.$$('su-toast .item').length).to.equal(1)
+    expect(component.$('su-toast .item').innerText).to.equal('hello')
 
     this.clock.tick(5010)
-    $('su-toast .item').length.should.equal(0)
+    expect(component.$('su-toast .item')).to.be.undefined
   })
 
   it('opens/closes toast', function () {
     mount()
-    $('su-toast .item').length.should.equal(0)
+    expect(component.$('su-toast .item')).to.be.undefined
 
-    app.suToast('hello')
-    tag.items[0].messages[0].should.equal('hello')
-    $('su-toast .item').length.should.equal(1)
-    $('su-toast .ui.message').length.should.equal(1)
+    component.suToast('hello')
+    expect(component.$('su-toast .item').innerText).to.equal('hello')
+    expect(component.$$('su-toast .item').length).to.equal(1)
+    expect(component.$('su-toast .ui.message')).to.not.be.undefined
 
-    $('su-toast .item:first .icon.close').click()
-    $('su-toast .item').length.should.equal(1)
-    $('su-toast .ui.message').length.should.equal(0)
+    fireEvent(component.$('su-toast .item .icon.close'), 'click')
+    expect(component.$$('su-toast .item').length).to.equal(1)
+    expect(component.$('su-toast .ui.message')).to.be.undefined
   })
 
   it('all option', function () {
     mount()
-    app.suToast({
+    component.suToast({
       title: 'riot',
       message: ['hello!', 'hello!!'],
       icon: 'check',
       class: 'info',
     })
-    tag.items[0].title.should.equal('riot')
-    tag.items[0].messages[0].should.equal('hello!')
-    tag.items[0].messages[1].should.equal('hello!!')
+    expect(component.$('su-toast .item .header').innerText.trim()).to.equal('riot')
+    expect(component.$$('su-toast .item p')[0].innerText).to.equal('hello!')
+    expect(component.$$('su-toast .item p')[1].innerText).to.equal('hello!!')
   })
 
   it('title only', function () {
     mount()
-    app.suToast({ title: 'riot' })
-    tag.items[0].title.should.equal('riot')
-    expect(tag.items[0].messages[0]).to.be.null
+    component.suToast({ title: 'riot' })
+    expect(component.$('su-toast .item .header').innerText.trim()).to.equal('riot')
+    expect(component.$('su-toast .item p').innerText).to.empty
   })
 
   it('messages only', function () {
     mount()
-    app.suToast({ message: ['hello!', 'hello!!'] })
-    expect(tag.items[0].title).to.be.null
-    tag.items[0].messages[0].should.equal('hello!')
-    tag.items[0].messages[1].should.equal('hello!!')
+    component.suToast({ message: ['hello!', 'hello!!'] })
+    expect(component.$('su-toast .item .header')).to.undefined
+    expect(component.$$('su-toast .item p')[0].innerText).to.equal('hello!')
+    expect(component.$$('su-toast .item p')[1].innerText).to.equal('hello!!')
   })
 
   it('none message', function () {
     mount()
-    app.suToast()
-    expect(tag.items[0].title).to.be.null
-    expect(tag.items[0].messages[0]).to.be.null
+    component.suToast()
+    expect(component.$('su-toast .item .header')).to.undefined
+    expect(component.$('su-toast .item p').innerText).to.empty
   })
 })
